requirements.txt
###### Requirements without Version Specifiers ######`
langchain
openai
faiss-cpu

ipykernel
python-dotenv

streamlit
streamlit-aggrid
streamlit-cropper
streamlit-cropperjs
streamlit-drawable-canvas==0.9.0 # 0.8.0
# svgpathtools
# svgwrite

# opencv-python
opencv-python-headless
pytesseract
pyocr
matplotlib
openpyxl
xlsxwriter

selenium
beautifulsoup4
lxml

Levenshtein
###### Requirements with Version Specifiers ######`
# docopt == 0.6.1             # Version Matching. Must be version 0.6.1
# keyring >= 4.1.1            # Minimum version 4.1.1
# coverage != 3.5             # Version Exclusion. Anything except version 3.5
# Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.*